requirements.txt
pandas==1.3.5
mysql-connector-python==8.0.27
psycopg2-binary==2.9.2